+++ resources/js/addon.js
@@ -10,41 +10,44 @@ const types = {
     heading: {
         tag: 'h',
         ext: 'heading',
+        cmd: 'btsToggleHeading',
         toggle: false,
-        cmd: 'btsToggleHeading'
     },
     paragraph: {
         tag: 'p',
         ext: 'paragraph',
+        cmd: 'btsToggleParagraph',
         toggle: false,
-        cmd: 'btsToggleParagraph'
     },
     span: {
         tag: 'span',
         ext: 'btsSpan',
+        cmd: 'btsToggleSpan',
         toggle: false,
-        cmd: 'btsToggleSpan'
     },
     link: {
         tag: 'a',
         ext: 'link',
+        cmd: 'btsToggleLink',
         toggle: true,
-        cmd: 'btsToggleLink'
     },
     bulletList: {
         tag: 'ul',
         ext: 'bulletList',
-        cmd: 'btsToggleBulletList'
+        cmd: 'btsToggleBulletList',
+        toggle: false,
     },
     orderedList: {
         tag: 'ol',
         ext: 'orderedList',
-        cmd: 'btsToggleOrderedList'
+        cmd: 'btsToggleOrderedList',
+        toggle: false,
     },
     div: {
         tag: 'div',
         ext: 'btsDiv',
-        cmd: 'btsToggleDiv'
+        cmd: 'btsToggleDiv',
+        toggle: false,
     },
 };
 
@@ -72,12 +75,24 @@ Statamic.booting(() => {
 
     Statamic.$bard.addExtension(() => Core.configure({ attr, styleTypes }));
     Statamic.$bard.addExtension(({ bard }) => {
+        const buttons = bard.buttons;
+        if (!buttons.find(button => button.name === 'bts_menu')) {
+            return;
+        }
         bard.buttons.forEach(button => {
-            if (bard.config.bts_menu.includes(button.name)) {
+            const menu = bard.config.bts_menu || [];
+            if (menu.includes(button.name)) {
                 button.visibleWhenActive = 'btsVoid';
             }
         });
-        return null;
+    });
+    Statamic.$bard.addExtension(({ bard }) => {
+        const nullify = ['h1', 'h2', 'h3', 'h4', 'h5', 'h6', 'ordererdlist', 'unordererdlist'];
+        bard.buttons.forEach(button => {
+            if (nullify.includes(button.name)) {
+                button.args.class = null;
+            }
+        });
     });
     Statamic.$bard.addExtension(() => Span);
     if (pro) {
@@ -102,21 +117,24 @@ Statamic.booting(() => {
                 args: args,
                 activeName: type.ext,
                 html: icon,
-                command: (editor, args) => editor.chain().focus()[type.cmd](args).run(),
+                isVisible: (editor) => editor.isActive(type.ext),
+                command: (editor, args) => editor.commands[type.cmd](args),
                 btsStyle: style,
             };
             buttons.splice(buttons.indexOf(key), 0, button(data));
         });
     });
-    // Statamic.$bard.buttons((buttons, button) => {
-    //     buttons.splice(buttons.indexOf('bts_menu'), 0, button({
-    //         name: 'bts_menu',
-    //         text: 'Style',
-    //         component: MenuButton,
-    //         html: `<svg viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg" xml:space="preserve" style="fill-rule:evenodd;clip-rule:evenodd;stroke-linejoin:round;stroke-miterlimit:2"><path fill="currentColor" d="M12 17.989V16h.532a.67.67 0 0 0 .633-.887L12.349 13H6.117l-.815 2.113a.67.67 0 0 0 .633.887H7v2H0v-2h.43a1.4 1.4 0 0 0 1.32-.935L7.257 0h3.952l4.768 13.042-.346.319c-1.407 1.317-1.454 1.899-1.909 4.069a.865.865 0 0 0-.016.258 1.074 1.074 0 0 0-.114-.006c-.648.007-1.166.12-1.592.307ZM7.114 10h4.238L9.093 3.801 7.114 10Z" style="fill-rule:nonzero"/><path fill="currentColor" d="M0 21.398c5.504.456 3.533-5.392 8.626-5.445l2.206 1.841c.549 6.645-7.579 8.127-10.832 3.604Zm16.878-8.538c1.713-2.687 7.016-11.698 7.016-11.698.423-.747-.515-1.528-1.17-.976 0 0-7.887 6.857-10.213 9.03-1.838 1.719-1.846 2.504-2.441 5.336l2.016 1.681c2.67-1.098 3.439-1.248 4.792-3.373Z" style="fill-rule:nonzero" transform="translate(7.766 7.766) scale(.6764)"/></svg>`,
-    //         btsConfig,
-    //     }));
-    // });
+    if (pro) {
+        Statamic.$bard.buttons((buttons, button) => {
+            buttons.splice(buttons.indexOf('bts_menu'), 0, button({
+                name: 'bts_menu',
+                text: 'Style',
+                component: MenuButton,
+                html: `<svg viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg" xml:space="preserve" style="fill-rule:evenodd;clip-rule:evenodd;stroke-linejoin:round;stroke-miterlimit:2"><path fill="currentColor" d="M12 17.989V16h.532a.67.67 0 0 0 .633-.887L12.349 13H6.117l-.815 2.113a.67.67 0 0 0 .633.887H7v2H0v-2h.43a1.4 1.4 0 0 0 1.32-.935L7.257 0h3.952l4.768 13.042-.346.319c-1.407 1.317-1.454 1.899-1.909 4.069a.865.865 0 0 0-.016.258 1.074 1.074 0 0 0-.114-.006c-.648.007-1.166.12-1.592.307ZM7.114 10h4.238L9.093 3.801 7.114 10Z" style="fill-rule:nonzero"/><path fill="currentColor" d="M0 21.398c5.504.456 3.533-5.392 8.626-5.445l2.206 1.841c.549 6.645-7.579 8.127-10.832 3.604Zm16.878-8.538c1.713-2.687 7.016-11.698 7.016-11.698.423-.747-.515-1.528-1.17-.976 0 0-7.887 6.857-10.213 9.03-1.838 1.719-1.846 2.504-2.441 5.336l2.016 1.681c2.67-1.098 3.439-1.248 4.792-3.373Z" style="fill-rule:nonzero" transform="translate(7.766 7.766) scale(.6764)"/></svg>`,
+                btsConfig,
+            }));
+        });
+    }
 
     // CSS
 
@@ -135,7 +153,7 @@ Statamic.booting(() => {
             .forEach(rule => css.push(rule.cssText.replaceAll(selector[0], selector[1])));
     }
 
-    Object.entries(styles).forEach(([, style]) => {
+    Object.entries(styles).forEach(([key, style]) => {
         if (!types[style.type]) {
             return;
         }