+++ src/ServiceProvider.php
@@ -96,26 +96,28 @@ class ServiceProvider extends AddonServiceProvider
             ]);
         }
 
-        // $buttons = collect($styles)
-        //     ->mapWithKeys(fn ($style, $key) => [$key => $style['name']])
-        //     ->merge([
-        //         'h1' => 'Heading 1',
-        //         'h2' => 'Heading 2',
-        //         'h3' => 'Heading 3',
-        //         'h4' => 'Heading 4',
-        //         'h5' => 'Heading 5',
-        //         'h6' => 'Heading 6',
-        //         'unorderedlist' => 'Unordered List',
-        //         'orderedlist' => 'Ordered List',
-        //     ]);
-        // Bard::appendConfigField('bts_menu', [
-        //     'display' => __('Style Menu'),
-        //     'instructions' => __('Which style options should be moved to the style menu'),
-        //     'type' => 'checkboxes',
-        //     'default' => [],
-        //     'multiple' => true,
-        //     'options' => $buttons,
-        // ]);
+        if ($pro) {
+            $buttons = collect($styles)
+                ->mapWithKeys(fn ($style, $key) => [$key => $style['name']])
+                ->merge([
+                    'h1' => 'Heading 1',
+                    'h2' => 'Heading 2',
+                    'h3' => 'Heading 3',
+                    'h4' => 'Heading 4',
+                    'h5' => 'Heading 5',
+                    'h6' => 'Heading 6',
+                    'unorderedlist' => 'Unordered List',
+                    'orderedlist' => 'Ordered List',
+                ]);
+            Bard::appendConfigField('bts_menu', [
+                'display' => __('Style Menu'),
+                'instructions' => __('Which style options should be moved to the style menu'),
+                'type' => 'checkboxes',
+                'default' => [],
+                'multiple' => true,
+                'options' => $buttons,
+            ]);
+        }
 
         return $this;
     }