+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 2.1.0 ()
 
--
+- [fix] Standard buttons activating for custom styles
 
 ## 2.0.1 (2023-01-31)